backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
pytest
numpy
google-generativeai